requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
python-dotenv==0.19.0
pymongo==4.6.1
SpeechRecognition==3.10.0
pydub==0.25.1
numpy==1.24.3
pyaudio==0.2.13
requests==2.31.0
wave==0.0.2
gitpython==3.1.40